Refactor: Extract canvas interactions and message sync logic into hooks
