+++ src/components/ui/canvas/hooks/index.ts
@@ -2,3 +2,5 @@ export * from './useCanvasComponentStore';
 export * from './useCanvasRehydration';
 export * from './useCanvasThreadReset';
 export * from './useCanvasEvents';
+export * from './useCanvasInteractions';
+export * from './useCanvasMessageSync';
